Order Api

+++ src/main/java/org/sang/foodorderingweb/model/Order.java
@@ -50,5 +50,5 @@ public class Order {
 
 	private int totalItem;
 
-	private float totalPrice;
+	private Long totalPrice;
 }
